Added giveaway page and the controller for it

+++ src/main/scala/model/Foodbank.scala
@@ -10,5 +10,5 @@ case class Foodbank(
                      operatingHour: String,
                      foodAvailable: String,
                      additionalInformation: String,
-                     owner: Float
+                     owner: Int
                    )
